Use Down on tvOS
